scripts: fallback to cap2hccapx if hcxpcapngtool fails

+++ scripts/pcapper.sh
@@ -28,9 +28,14 @@ for file in $PCAP_DIR/*.pcap; do
     
     hcxpcapngtool --all -o "$STORE_DIR/capture.hccapx" "$STORE_DIR/$JUST_NAME"
     # if this fails we weren't able to capture any crackable materials 
-    # so wipe out the data 
+    # so wipe out the data, however note that this might not be entirely accurate
     if [[ ! -f "$STORE_DIR/capture.hccapx" ]]; then
-        rm -rf "$STORE_DIR"
+        # try cap2hccapx although this will likely also not work
+        cap2hccapx "$STORE_DIR/$JUST_NAME" "$STORE_DIR/capture.hccapx"
+        if [[ $(du -hs "$STORE_DIR/capture.hccapx" | awk '{print $1}') == 0 ]]; then
+            echo "pcap conversion failed for $JUST_NAME_NO_EXT"
+            rm -rf "$STORE_DIR"
+        fi
     fi
 
 done
